Move stashbox package under pkg
