+++ pkg/stashbox/draft.go
