+++ pkg/stashbox/graphql/generated_client.go
