+++ pkg/stashbox/graphql/generated_models.go
