+++ pkg/stashbox/models.go
